booking_services/requirements.txt
Flask
flask_sqlalchemy
flask_cors
psycopg2-binary
requests
pika
python-dotenv